using copious-hosts-manager as template make new project
This is for managing the software with minimal interfaces.
For example, detailed work is being done in git-cola at this moment.
But, the repo list, can give fast access to things, fire vscode, etc.

One reason, is that in moving to a new laptop, lots of things didn't
copy just right. Need to sync with github, etc.